refactor : move ErrorBoundary folder to components folder
