+++ openmetadata-ui/src/main/resources/ui/src/App.tsx
@@ -30,12 +30,12 @@ import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.min.css';
 import { AuthProvider } from './authentication/auth-provider/AuthProvider';
 import Appbar from './components/app-bar/Appbar';
+import ErrorBoundry from './components/ErrorBoundry/ErrorBoundry';
 import GlobalSearchProvider from './components/GlobalSearchProvider/GlobalSearchProvider';
 import PermissionProvider from './components/PermissionProvider/PermissionProvider';
 import WebSocketProvider from './components/web-scoket/web-scoket.provider';
 import WebAnalyticsProvider from './components/WebAnalytics/WebAnalyticsProvider';
 import { TOAST_OPTIONS } from './constants/Toasts.constants';
-import ErrorBoundry from './ErrorBoundry/ErrorBoundry';
 import AppRouter from './router/AppRouter';
 
 const App: FunctionComponent = () => {
