+++ openmetadata-ui/src/main/resources/ui/src/components/ErrorBoundry/ErrorBoundry.tsx
@@ -11,30 +11,28 @@
  *  limitations under the License.
  */
 
-import { Button, Result } from 'antd';
 import React from 'react';
-import { FallbackProps } from 'react-error-boundary';
-import { ERROR500 } from '../constants/constants';
+import { ErrorBoundary } from 'react-error-boundary';
+import { useHistory } from 'react-router-dom';
+import { ROUTES } from '../../constants/constants';
+import ErrorFallback from './ErrorFallback';
 
-const ErrorFallback: React.FC<FallbackProps> = ({
-  error,
-  resetErrorBoundary,
-}) => {
+interface Props {
+  children: React.ReactNode;
+}
+
+const ErrorBoundry: React.FC<Props> = ({ children }) => {
+  const history = useHistory();
+
+  const onErrorReset = () => {
+    history.push(ROUTES.HOME);
+  };
+
   return (
-    <Result
-      extra={
-        <Button
-          className="ant-btn-primary-custom"
-          type="primary"
-          onClick={resetErrorBoundary}>
-          Home
-        </Button>
-      }
-      status="404"
-      subTitle={error.message}
-      title={ERROR500}
-    />
+    <ErrorBoundary FallbackComponent={ErrorFallback} onReset={onErrorReset}>
+      {children}
+    </ErrorBoundary>
   );
 };
 
-export default ErrorFallback;
+export default ErrorBoundry;
