+++ openmetadata-ui/src/main/resources/ui/src/components/ErrorBoundry/ErrorFallback.tsx
@@ -11,28 +11,30 @@
  *  limitations under the License.
  */
 
+import { Button, Result } from 'antd';
 import React from 'react';
-import { ErrorBoundary } from 'react-error-boundary';
-import { useHistory } from 'react-router-dom';
-import { ROUTES } from '../constants/constants';
-import ErrorFallback from './ErrorFallback';
+import { FallbackProps } from 'react-error-boundary';
+import { ERROR500 } from '../../constants/constants';
 
-interface Props {
-  children: React.ReactNode;
-}
-
-const ErrorBoundry: React.FC<Props> = ({ children }) => {
-  const history = useHistory();
-
-  const onErrorReset = () => {
-    history.push(ROUTES.HOME);
-  };
-
+const ErrorFallback: React.FC<FallbackProps> = ({
+  error,
+  resetErrorBoundary,
+}) => {
   return (
-    <ErrorBoundary FallbackComponent={ErrorFallback} onReset={onErrorReset}>
-      {children}
-    </ErrorBoundary>
+    <Result
+      extra={
+        <Button
+          className="ant-btn-primary-custom"
+          type="primary"
+          onClick={resetErrorBoundary}>
+          Home
+        </Button>
+      }
+      status="404"
+      subTitle={error.message}
+      title={ERROR500}
+    />
   );
 };
 
-export default ErrorBoundry;
+export default ErrorFallback;
